requirements.txt
langchain_google_genai
os
streamlit
dotenv
langchain_community.document_loaders
langchain_google_genai
langchain_chroma
langchain.chains
langchain.chains.combine_documents
langchain_core.prompts